+++ src/components/icon/Icon.vue
@@ -213,6 +213,14 @@ export default {
                     key: 'wetroom',
                     value: 'level-entry-shower',
                 },
+                {
+                    key: 'x-twitter',
+                    value: 'x-twitter fa-brands',
+                },
+                {
+                    key: 'linkedin',
+                    value: 'linkedin-in fa-brands',
+                },
             ],
         };
     },